% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{getFRgivenG}
\alias{getFRgivenG}
\title{Posterior probabilities of FR given G}
\usage{
getFRgivenG(G, logistic.model, lambda)
}
\arguments{
\item{G}{Genomic features}

\item{logistic.model}{Logistic regression model with current estimate of beta}

\item{lambda}{Selected lambda}
}
\value{
P(FR | G)
}
\description{
\code{getFRgivenG} computes posterior probabilities of FR (functionality of
        regulatory variant) given G (genomic annotations) and current estimate
        of beta (parameters between FR and G).
}
\examples{
dataInput <- getData(filename=system.file("extdata", "simulation_RIVER.gz",
        package = "RIVERpkg"), ZscoreThrd=1.5)
G = scale(t(Biobase::exprs(dataInput))) # genomic features (G)
E = as.vector(as.numeric(unlist(dataInput$Outlier))-1) # outlier status (E)
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(G, E, lambda=costs, family="binomial",
        alpha = 0, nfolds=10)
p.FR.givenG = getFRgivenG(G, logistic.model=cv.all.ll$glmnet.fit,
        lambda=cv.all.ll$lambda.min)

}
\seealso{
\code{\link{predict}}
}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
