% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{mleTheta}
\alias{mleTheta}
\title{Maximum likelihoood estimate of theta.}
\usage{
mleTheta(E, FR, pseudocount)
}
\arguments{
\item{E}{Binary values of outlier status.}

\item{FR}{Soft-assignments of FR from E-step}

\item{pseudocount}{Pseudo count.}
}
\value{
MLE of theta
}
\description{
\code{mleTheta} computes maximum likelihoood estimate of theta (parameters
        between FR (functionality of regulatory variant) and E (outlier
        status); Naive-Bayes).
}
\examples{
dataInput <- getData(filename=system.file("extdata", "simulation_RIVER.gz",
        package = "RIVERpkg"), ZscoreThrd=1.5)
G = scale(t(Biobase::exprs(dataInput))) # genomic features (G)
E = as.vector(as.numeric(unlist(dataInput$Outlier))-1) # outlier status (E)
theta_init=matrix(c(.99, .01, .3, .7), nrow=2)
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(G, E, lambda=costs, family="binomial",
        alpha = 0, nfolds=10)
p.FR.givenG = getFRgivenG(G, cv.all.ll$glmnet.fit, cv.all.ll$lambda.min)
posteriors = getFRPosteriors(E, p.FR.givenG, theta=theta_init)
theta.cur = mleTheta(E, FR=posteriors$posterior, pseudoc=50)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
