% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcPLOTs.R
\name{plotAUC}
\alias{plotAUC}
\title{Draw AUC curves from both RIVER and GAM.}
\usage{
plotAUC(rocstat)
}
\arguments{
\item{rocstat}{Output of \code{evaRIVER}, which provides AUC values from both
RIVER and GAM models and p-value between two AUCs.}
}
\value{
AUC figure
}
\description{
\code{plotAUC} plots AUC curves from both RIVER and GAM (genomic annotation
        model).
}
\examples{
G <- simulated_features
E <- simulated_outliers
P <- simulated_N2pairs
rocstat <- evaRIVER(G, E, P, pseudoc=50, theta_init=matrix(c(.99, .01, .3, .7),
            nrow=2), costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4), verbose=FALSE)
plotAUC(rocstat)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}

