% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcPLOTs.R
\name{plotPosteriors}
\alias{plotPosteriors}
\title{Draw scatter plots of posterior probabilities from both RIVER GAM in terms
        of outlier status.}
\usage{
plotPosteriors(postprobs, outliers)
}
\arguments{
\item{postprobs}{Output of \code{evaRIVER}, which provides test posterior
probabilities from both RIVER and GAM for all instances.}

\item{outliers}{Outlier status of examples}
}
\value{
A figure of posteriors from RIVER (y-axis) and GAM (x-axis) models
        for ouliters and non-outliers separately
}
\description{
\code{plotPosteriors} draws scatter plots of posterior probabilities from
        both RIVER GAM (genomic annotation model) in terms of outlier
        status.
}
\examples{
G <- simulated_features
E <- simulated_outliers
outRIVER <- appRIVER(G, E, pseudoc=50, theta_init=matrix(c(.99, .01, .3, .7),
        nrow=2), costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4), verbose=TRUE)
plotPosteriors(outRIVER, E)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}

