% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_norway.R
\docType{data}
\name{river_norway}
\alias{river_norway}
\title{Norwegian rivers}
\format{A rda file}
\source{
\href{http://www.nve.no}{NVE}
}
\description{
River shapefile convert into R format.
The original dataset is available at http://nedlasting.nve.no/gis/
}
\references{
the Norwegian Water Resources and Energy Directorate (NVE)
}
\keyword{datasets}

