% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_outliers}
\alias{simulated_outliers}
\title{Binary values of gene expression outlier status.}
\format{A data frame with 6122 rows and 3 columns including subject ID, gene
        name, and outlier status:
  \describe{
  \item{SubjectID}{subject IDs of 115 distinct individuals}
  \item{GeneName}{1631 distinct gene names}
  \item{Outlier}{binary values of outlier status}}}
\source{
\url{https://github.com/ipw012/RIVERpkg}
}
\usage{
simulated_outliers
}
\value{
A simulated outlier data in a data.frame format (6122 x 3)
}
\description{
A simulated dataset  having binary values of gene expression outlier status
        for 6122 instances. Note that simulated_features and simulated_outliers
        are matched in the same order.
}
\keyword{datasets}

