% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{testPosteriors}
\alias{testPosteriors}
\title{Test posterior probabilities of FR given G and E}
\usage{
testPosteriors(G, E, em.results)
}
\arguments{
\item{G}{Genomic features}

\item{E}{Binary values of outlier status.}

\item{em.results}{Estimated parameters including beta and theta via EM and
selected lambdas}
}
\value{
P(FR | G, E, beta, theta) and probable status of FR.
}
\description{
\code{testPosteriors} computes posterior probabilities of FR (functionality
        of regulatory variant) given G (genomic annotations) and E (outlier
        status) with estimate of beta (parameters between FR and G) and
        theta (parameters between FR and E).
}
\examples{
G <- simulated_features
E <- simulated_outliers
g_all <- scale(as.matrix(G[,3:ncol(G)]))
E_disc = as.vector(as.numeric(E[,"Outlier"]))
theta_init=matrix(c(.99, .01, .3, .7), nrow=2)
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(g_all, E_disc, lambda=costs, family="binomial",
        alpha = 0, nfolds=10)
em.all.res <- integratedEM(g_all, E_disc, cv.all.ll$lambda.min,
cv.all.ll$glmnet.fit, pseudoc=50, theta_init, costs, verbose=FALSE)
train.post = testPosteriors(g_all, E_disc, em.all.res)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
\seealso{
\code{\link{getFRgivenG}} and \code{\link{getFRPosteriors}}
}

