% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaRIVER.R
\name{evaRIVER}
\alias{evaRIVER}
\title{Evaluation of RIVER}
\usage{
evaRIVER(G, E, pseudoc = 50, theta_init = matrix(c(0.99, 0.01, 0.3, 0.7),
  nrow = 2), costs = c(100, 10, 1, 0.1, 0.01, 0.001, 1e-04),
  verbose = FALSE)
}
\arguments{
\item{G}{Genomic features in the data frame which contains subject IDs,
gene symbols, and P number of genomic fetures of each instance
per row.}

\item{E}{Expression outliers in the data frame which contains subject IDs,
gene symbols, and outlier status of each instance per row.}

\item{pseudoc}{Pseudo count.}

\item{theta_init}{Initial values of theta.}

\item{costs}{Candidate penalty parameter values for L2-regularized logistic
regression.}

\item{verbose}{Logical option for showing extra information on progress.}
}
\value{
An object of class RIVER which include two AUC values from
        RIVER and GAM and corresponding P-value of comparing the two AUC
        values.
}
\description{
\code{evaRIVER} trains RIVER by holding out a list of individual and gene
        pairs having same rare variants for evaluation, computes test
        posterior probabilities of FR for 1st individual, and compares
        them with outlier status of 2nd individual from the list.
}
\section{Warning}{
 A vector of candidate penalty values make \code{glmnet}
        faster than to input a single penalty value
}
\examples{
G <- simulated_features
E <- simulated_outliers
rocSTAT <- evaRIVER(G, E, verbose=TRUE)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}, \code{\link{predict}},
        \code{\link{integratedEM}}, and \code{\link{testPosteriors}}
}

