% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{getFRgivenG}
\alias{getFRgivenG}
\title{Posterior probabilities of FR given G}
\usage{
getFRgivenG(G, logistic.model, lambda)
}
\arguments{
\item{G}{Genomic features}

\item{logistic.model}{Logistic regression model with current estimate of beta}

\item{lambda}{Selected lambda}
}
\value{
P(FR | G)
}
\description{
\code{getFRgivenG} computes posterior probabilities of FR (functionality of
        regulatory variant) given G (genomic annotations) and current estimate
        of beta (parameters between FR and G).
}
\examples{
G <- simulated_features
E <- simulated_outliers
g_all <- scale(as.matrix(G[,3:ncol(G)]))
E_disc = as.vector(as.numeric(E[,"Outlier"]))
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(g_all, E_disc, lambda=costs, family="binomial",
        alpha = 0, nfolds=10)
p.FR.givenG = getFRgivenG(g_all, cv.all.ll$glmnet.fit, cv.all.ll$lambda.min)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
\seealso{
\code{\link{predict}}
}

