% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{getFRPosteriors}
\alias{getFRPosteriors}
\title{Posterior probabilities of FR given G and E.}
\usage{
getFRPosteriors(E, p.FR.givenG, theta)
}
\arguments{
\item{E}{Binary values of outlier status.}

\item{p.FR.givenG}{P(FR | G, beta) from \code{getFRgivenG}.}

\item{theta}{Current estimate of theta.}
}
\value{
P(FR | G, E, beta, theta) and probable status of FR.
}
\description{
\code{getFRPosteriors} computes posterior probabilities of FR (functionality
        of regulatory variant) given G (genomic annotations) and E (outlier
        status) with current estimate of beta (parameters between FR and G)
        and theta (parameters between FR and E).
}
\examples{
dataInput <- getData(filename=system.file("extdata", "simulation_RIVER.gz",
        package = "RIVERpkg"), ZscoreThrd=1.5)
G = scale(t(Biobase::exprs(dataInput))) # genomic features (G)
E = as.vector(as.numeric(unlist(dataInput$Outlier))-1) # outlier status (E)
theta_init=matrix(c(.99, .01, .3, .7), nrow=2)
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(G, E, lambda=costs, family="binomial",
        alpha=0, nfolds=10)
p.FR.givenG = getFRgivenG(G, cv.all.ll$glmnet.fit, cv.all.ll$lambda.min)
posteriors = getFRPosteriors(E, p.FR.givenG, theta=theta_init)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
