% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{integratedEM}
\alias{integratedEM}
\title{An iterative expectation-maximization algorithm for RIVER}
\usage{
integratedEM(G, E, lambda, logistic.init, pseudoc, theta.init, costs,
  verbose = FALSE)
}
\arguments{
\item{G}{Genomic features.}

\item{E}{Binary values of outlier status.}

\item{lambda}{Selected lambda.}

\item{logistic.init}{Smart initialization of beta (parameters between
FR and G) from estimate of beta with E via multivariate logistic
regression.}

\item{pseudoc}{Pseudo count.}

\item{theta.init}{Initial theta (parameters between FR (functionality
of regulatory variant) and E).}

\item{costs}{Candidate penalty parameter values for L2-regularization
within logistic regression.}

\item{verbose}{Logical option for showing extra information on progress.}
}
\value{
Best estimate of beta and theta, final multivariate logistic
        regression model, and posterior probabilities of FR.
}
\description{
\code{integratedEM} iteratively executes e-step and m-step until it
        converges. This is a main function of RIVER.
}
\examples{
dataInput <- getData(filename=system.file("extdata", "simulation_RIVER.gz",
        package = "RIVERpkg"), ZscoreThrd=1.5)
G = scale(t(Biobase::exprs(dataInput))) # genomic features (G)
E = as.vector(as.numeric(unlist(dataInput$Outlier))-1) # outlier status (E)
theta_init=matrix(c(.99, .01, .3, .7), nrow=2)
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(G, E, lambda=costs, family="binomial",
        alpha = 0, nfolds=10)
em.all.res <- integratedEM(G, E, lambda=cv.all.ll$lambda.min,
        logistic.init=cv.all.ll$glmnet.fit, pseudoc=50, theta=theta_init,
        costs, verbose=FALSE)

}
\seealso{
\code{\link{getFRgivenG}}, \code{\link{getFRPosteriors}},
        \code{\link{mleTheta}}, \code{\link{mleBeta}},
        \code{\link[glmnet]{cv.glmnet}},
        and \url{https://github.com/ipw012/RIVERpkg}
}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
