% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{mleBeta}
\alias{mleBeta}
\title{Maximum likelihoood estimate of beta.}
\usage{
mleBeta(G, FR, costs)
}
\arguments{
\item{G}{Genomic features}

\item{FR}{Soft-assignments of FR from E-step}

\item{costs}{Candidate penalty parameter values for L2-regularization within
logistic regression.}
}
\value{
MLE of beta
}
\description{
\code{mleBeta} computes maximum likelihoood estimate of beta (parameters
        between FR (functionality of regulatory variant) and G (genomic
        annotations); multivariate logistic regression).
}
\section{Warning}{
 To input a vector of candidate penalty values makes
        \code{glmnet} faster than to input a single penalty value
}
\examples{
G <- simulated_features
E <- simulated_outliers
g_all <- scale(as.matrix(G[,3:ncol(G)]))
E_disc = as.vector(as.numeric(E[,"Outlier"]))
theta_init=matrix(c(.99, .01, .3, .7), nrow=2)
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(g_all, E_disc, lambda=costs, family="binomial",
        alpha=0, nfolds=10)
p.FR.givenG = getFRgivenG(g_all, cv.all.ll$glmnet.fit, cv.all.ll$lambda.min)
posteriors = getFRPosteriors(E_disc, p.FR.givenG, theta=theta_init)
logistic.cur = mleBeta(g_all, posteriors$posterior, costs)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
\seealso{
\code{\link[glmnet]{glmnet}}
}

