% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{mleTheta}
\alias{mleTheta}
\title{Maximum likelihoood estimate of theta.}
\usage{
mleTheta(E, FR, pseudocount)
}
\arguments{
\item{E}{Binary values of outlier status.}

\item{FR}{Soft-assignments of FR from E-step}

\item{pseudocount}{Pseudo count.}
}
\value{
MLE of theta
}
\description{
\code{mleTheta} computes maximum likelihoood estimate of theta (parameters
        between FR (functionality of regulatory variant) and E (outlier
        status); Naive-Bayes).
}
\examples{
G <- simulated_features
E <- simulated_outliers
g_all <- scale(as.matrix(G[,3:ncol(G)]))
E_disc = as.vector(as.numeric(E[,"Outlier"]))
theta_init=matrix(c(.99, .01, .3, .7), nrow=2)
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(g_all, E_disc, lambda=costs, family="binomial",
        alpha = 0, nfolds=10)
p.FR.givenG = getFRgivenG(g_all, cv.all.ll$glmnet.fit, cv.all.ll$lambda.min)
posteriors = getFRPosteriors(E_disc, p.FR.givenG, theta=theta_init)
theta.cur = mleTheta(E_disc, posteriors$posterior, pseudoc=50)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}

