% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcPLOTs.R
\name{plotAUC}
\alias{plotAUC}
\title{Draw AUC curves from both RIVER and GAM.}
\usage{
plotAUC(evaROC)
}
\arguments{
\item{evaROC}{Output of \code{evaRIVER}, an S4 object of class evaRIVER which
include two AUC values from RIVER and GAM, computed specificities and
sensitivities from two models, and P-value of comparing the two AUC
values.}
}
\value{
AUC figure
}
\description{
\code{plotAUC} plots AUC curves from both RIVER and GAM (genomic annotation
        model).
}
\examples{
data <- getData(filename=system.file("extdata", "simulation_RIVER.gz",
        package = "RIVERpkg"), ZscoreThrd=1.5)
evaROC <- evaRIVER(data, verbose=TRUE)
plotAUC(evaROC)

}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
