% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRIVER.R
\name{testPosteriors}
\alias{testPosteriors}
\title{Test posterior probabilities of FR given G and E}
\usage{
testPosteriors(G, E, em.results)
}
\arguments{
\item{G}{Genomic features}

\item{E}{Binary values of outlier status.}

\item{em.results}{Estimated parameters including beta and theta via EM and
selected lambdas}
}
\value{
P(FR | G, E, beta, theta) and probable status of FR.
}
\description{
\code{testPosteriors} computes posterior probabilities of FR (functionality
        of regulatory variant) given G (genomic annotations) and E (outlier
        status) with estimate of beta (parameters between FR and G) and
        theta (parameters between FR and E).
}
\examples{
dataInput <- getData(filename=system.file("extdata", "simulation_RIVER.gz",
        package = "RIVERpkg"), ZscoreThrd=1.5)
G = scale(t(Biobase::exprs(dataInput))) # genomic features (G)
E = as.vector(as.numeric(unlist(dataInput$Outlier))-1) # outlier status (E)
theta_init=matrix(c(.99, .01, .3, .7), nrow=2)
costs=c(100, 10, 1, .1, .01, 1e-3, 1e-4)
cv.all.ll = glmnet::cv.glmnet(G, E, lambda=costs, family="binomial",
        alpha = 0, nfolds=10)
em.all.res <- integratedEM(G, E, cv.all.ll$lambda.min, cv.all.ll$glmnet.fit,
        pseudoc=50, theta_init, costs, verbose=FALSE)
train.post = testPosteriors(G, E, em.results=em.all.res)

}
\seealso{
\code{\link{getFRgivenG}} and \code{\link{getFRPosteriors}}
}
\author{
Yungil Kim, \email{ipw012@gmail.com}
}
